requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
alembic==1.13.1
sentry-sdk==1.39.1
python-logging-loki==0.3.1
